// crates/madari-tv/web/src/client/types.gen.ts
export type AddonSummary = {
    allow_local: boolean;
    enabled: boolean;
    installation_id: string;
    manifest: Manifest;
};

export type Catalog = {
    extra?: Array<Extra>;
    id: string;
    name?: string | null;
    type: string;
    [key: string]: unknown;
};
// ...
export type Manifest = {
    addonCatalogs?: Array<Catalog>;
    behaviorHints?: {
        [key: string]: unknown;
    };
    catalogs?: Array<Catalog>;
    id: string;
    idPrefixes?: Array<string> | null;
    name: string;
    resources: Array<ResourceDeclaration>;
    types: Array<string>;
    version: string;
    [key: string]: unknown;
};
// ...
export type ProfileDto = {
    avatar?: string | null;
    id: string;
    kids: boolean;
    name: string;
    pin_protected: boolean;
};
// ...
export type ResourceDeclaration = string | {
    idPrefixes?: Array<string> | null;
    name: string;
    types?: Array<string>;
    [key: string]: unknown;
};